+++ lib/tty/option/deep_dup.rb
@@ -2,46 +2,70 @@
 
 module TTY
   module Option
+    # Responsible for deep copying an object
+    #
+    # @api private
     module DeepDup
       NONDUPLICATABLE = [
         Symbol, TrueClass, FalseClass, NilClass, Numeric, Method, UnboundMethod
       ].freeze
 
-      # Duplicate an object making a deep copy
+      # Deep copy an object
+      #
+      # @example
+      #   DeepDeup.deep_dup({foo: {bar: [1, 2]}})
       #
       # @param [Object] object
+      #   the object to deep copy
+      # @param [Hash] cache
+      #   the cache of copied objects
+      #
+      # @return [Object]
       #
       # @api public
-      def self.deep_dup(object)
-        case object
-        when *NONDUPLICATABLE then object
-        when Hash   then deep_dup_hash(object)
-        when Array  then deep_dup_array(object)
-        else object.dup
-        end
+      def self.deep_dup(object, cache = {})
+        cache[object.object_id] ||=
+          case object
+          when *NONDUPLICATABLE then object
+          when Array then deep_dup_array(object, cache)
+          when Hash  then deep_dup_hash(object, cache)
+          else object.dup
+          end
       end
 
-      # A deep copy of hash
-      #
-      # @param [Hash] object
-      #
-      # @api private
-      def self.deep_dup_hash(object)
-        object.each_with_object({}) do |(key, val), new_hash|
-          new_hash[deep_dup(key)] = deep_dup(val)
-        end
-      end
-
-      # A deep copy of array
+      # Deep copy an array
       #
       # @param [Array] object
+      #   the array object to deep copy
+      # @param [Hash] cache
+      #   the cache of copied objects
+      #
+      # @return [Array]
       #
       # @api private
-      def self.deep_dup_array(object)
+      def self.deep_dup_array(object, cache)
         object.each_with_object([]) do |val, new_array|
-          new_array << deep_dup(val)
+          new_array << deep_dup(val, cache)
         end
       end
+      private_class_method :deep_dup_array
+
+      # Deep copy a hash
+      #
+      # @param [Hash] object
+      #   the hash object to deep copy
+      # @param [Hash] cache
+      #   the cache of copied objects
+      #
+      # @return [Hash]
+      #
+      # @api private
+      def self.deep_dup_hash(object, cache)
+        object.each_with_object({}) do |(key, val), new_hash|
+          new_hash[deep_dup(key, cache)] = deep_dup(val, cache)
+        end
+      end
+      private_class_method :deep_dup_hash
     end # DeepDup
   end # Option
 end # TTY