+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Change regexp conversion to only work with strings and regexp objects
 * Change uri conversion to work with uri objects
 * Change deep cloning to stop making copies of unbound methods
+* Change deep cloning to copy objects with the same identity only once
 
 ### Fixed
 * Fix argument check against permitted values to allow nil value when optional